// utils/types.ts
export type CreateUserParams = {
    username: string;
    firstName: string;
    lastName: string;
    password: string;
  };